Consultancy Agreement
I took a sample agreement from [ClearTax](https://cleartax.in/s/consultancy-agreement-format-download) but the format was not up to the latest standards. With the help of [Mr. Aditya Gupta](https://github.com/ag93999) and his friend, Ms. Surbhi, I modified the document as per the latest standard. Furthermore, I have tried to ensure that in this agreement, both parties, the Consultant & the Company are protected and bound by NDA. While the nature of agreement may vary in your scenario, feel free to modify and use document as per your need.